pipeline/requirements.txt
groq==0.9.0
cohere==5.5.0
supabase==2.7.0
requests==2.32.0
python-dotenv==1.0.0
tqdm==4.66.0
google-generativeai>=0.5.0
